Tidy up program generation.
