+++ src/main/kotlin/nz/co/jedsimson/lgp/lib/BaseProgramGenerator.kt
@@ -7,16 +7,12 @@ import nz.co.jedsimson.lgp.core.program.instructions.Instruction
 import nz.co.jedsimson.lgp.core.program.instructions.RegisterIndex
 import nz.co.jedsimson.lgp.core.program.Program
 import nz.co.jedsimson.lgp.core.program.ProgramGenerator
-import nz.co.jedsimson.lgp.core.evolution.operators.choice
 import nz.co.jedsimson.lgp.core.evolution.operators.randInt
-import nz.co.jedsimson.lgp.core.program.registers.RegisterType
 import nz.co.jedsimson.lgp.core.modules.ModuleInformation
 import nz.co.jedsimson.lgp.core.program.Output
 
 /**
- * A ``ProgramGenerator`` implementation that provides effective ``BaseProgram`` instances.
- *
- * This generator only creates programs which are "effective".
+ * A ``ProgramGenerator`` implementation that provides random ``BaseProgram`` instances.
  *
  * @property sentinelTrueValue A value that should be considered as boolean "true".
  * @property outputRegisterIndices A collection of indices that should be considered as the program output registers.
@@ -28,63 +24,35 @@ class BaseProgramGenerator<TProgram, TOutput : Output<TProgram>>(
         val outputRegisterIndices: List<RegisterIndex>,
         val outputResolver: (BaseProgram<TProgram, TOutput>) -> TOutput
 ) : ProgramGenerator<TProgram, TOutput>(
-        environment,
-        instructionGenerator = environment.registeredModule(CoreModuleType.InstructionGenerator)
+    environment,
+    instructionGenerator = environment.registeredModule(CoreModuleType.InstructionGenerator)
 ) {
-
     private val random = this.environment.randomState
 
     override fun generateProgram(): Program<TProgram, TOutput> {
         val length = this.random.randInt(
-                this.environment.configuration.initialMinimumProgramLength,
-                this.environment.configuration.initialMaximumProgramLength
+            this.environment.configuration.initialMinimumProgramLength,
+            this.environment.configuration.initialMaximumProgramLength
         )
 
-        val branchesUsed = this.environment.operations.any { op -> op is BranchOperation<TProgram> }
-        val output = this.environment.registerSet.calculationRegisters.start
-
         val instructions = mutableListOf<Instruction<TProgram>>()
-        // TODO: Use set instead of list.
-        val effectiveRegisters = mutableListOf<RegisterIndex>()
 
-        // Add first instruction
-        instructions.add(this.instructionGenerator.generateInstruction())
+        val branchesUsed = this.environment.operations.any { op -> op is BranchOperation<TProgram> }
+        val branchInitialisationRate = this.environment.configuration.branchInitialisationRate
 
-        // Construct effective instructions
-        for (i in 2..length) {
-            if (instructions.first().operation !is BranchOperation<TProgram>) {
-                effectiveRegisters.remove(instructions.first().destination)
-            }
-
-            // Add any operand registers that are not already known to be
-            // effective and are calculation registers.
-            instructions.first().operands.filter { operand ->
-                operand !in effectiveRegisters &&
-                        (this.environment.registerSet.registerType(operand) == RegisterType.Calculation)
-            }.forEach { operand -> effectiveRegisters.add(operand) }
-
-            if (effectiveRegisters.isEmpty()) {
-                effectiveRegisters.add(output)
-            }
-
-            if (branchesUsed && random.nextDouble() < this.environment.configuration.branchInitialisationRate) {
-                val instr = this.instructionGenerator.next().first { instruction ->
-                    instruction.operation is BranchOperation<TProgram>
+        for (i in 1..length) {
+            val instruction = when {
+                branchesUsed && random.nextDouble() < branchInitialisationRate -> {
+                    this.instructionGenerator.next().first { instruction ->
+                        instruction.operation is BranchOperation<TProgram>
+                    }
                 }
-
-                assert(instr.operation is BranchOperation<TProgram>)
-
-                instructions.add(0, instr)
-            } else {
-                // Get a random instruction and make it effective by
-                // using one of the registers marked as effective.
-                val instr = this.instructionGenerator.generateInstruction()
-
-                instr.destination = random.choice(effectiveRegisters)
-
-                // And add it to the program
-                instructions.add(0, instr)
+                else -> {
+                    this.instructionGenerator.generateInstruction()
+                }
             }
+
+            instructions.add(instruction)
         }
 
         // Each program gets its own copy of the register set
@@ -98,6 +66,6 @@ class BaseProgramGenerator<TProgram, TOutput : Output<TProgram>>(
     }
 
     override val information = ModuleInformation(
-        description = "A simple program generator."
+        description = "A ProgramGenerator implementation that provides random BaseProgram instances."
     )
 }
